refactor(#3038): VoiceBargeInRuntime S7 — fold agent-voice routing helpers into routing child
S7 scope: extract the agent-voice routing helper block that the S6 codex review
named as the precise S7 candidate (the S6 closing assessment was corrected for
overstating closure). This is the free-function cluster at root lines ~398-494 —
`agent_voice_matches_channel`, `agent_voice_channel`, `agent_voice_background_channel`,
the reverse/source-channel resolvers (`agent_voice_background_channel_for`,
`agent_voice_channel_for_background`, `agent_voice_source_channel_for_background`,
`effective_voice_source_channel`), and `agent_text_channel_matches`. They are
config-only pure resolvers with no struct-field dependency, and their dominant
consumer is already the `voice_barge_in/routing.rs` child (its
`resolve_effective_foreground_config`, `resolve_voice_background_channel_for_source`,
`voice_channel_for_background`, and `resolve_voice_turn_target` all call into the
block). routing.rs is 383 prod LoC; +104 lands it at 487, well under the 1000-LoC
giant threshold, so it is the natural destination (no new child needed).

Cluster boundary: moved all eight free functions verbatim into
`src/services/discord/voice_barge_in/routing.rs` as module-level (non-impl) items.
The only moved-symbol edits are visibility-only:

- All eight: private `fn` -> `pub(super) fn`. Three signatures
  (`agent_voice_matches_channel`, `agent_text_channel_matches`, and
  `agent_voice_background_channel`'s neighbours) rustfmt-rewrap because the
  `pub(super) ` prefix pushes the parameter list past the 100-col width; the
  bodies are byte-identical (normalized SHA below).

Child-boundary import edits (the only non-visibility edits, allowed by the
S-series discipline):

- Root prod re-import `use routing::{agent_voice_background_channel,
  agent_voice_matches_channel, effective_voice_source_channel};` so the three
  prod call sites resolve unqualified: `agent_voice_matches_channel`
  (try_handle_voice_channel_text_reply, process_completed_utterance),
  `agent_voice_background_channel` (try_handle_voice_channel_text_reply), and
  `effective_voice_source_channel` (handle_processing_transcript).
- Test-only re-import inside `mod tests`:
  `use super::routing::{agent_voice_background_channel_for,
  agent_voice_channel_for_background, agent_voice_source_channel_for_background};`
  for the four characterization tests; the prod-imported three reach the tests
  through the existing `use super::*`.
- `agent_voice_channel` and `agent_text_channel_matches` have no root caller
  (only intra-child use), so they are intentionally NOT re-imported into root —
  keeps the root import warning-clean under `cargo clippy --lib -- -D warnings`.

Within routing.rs the in-cluster calls (`agent_voice_matches_channel` from
`agent_voice_background_channel_for`, etc.) now resolve as same-module siblings;
the moved bodies reference only `crate::config::{AgentDef, Config}` (fully
qualified) and `ChannelId` (via the child's `use super::*`), so no body text
changed.

No await point, lock scope, spawn, map-entry, cancellation, channel-ordering, or
queue-mutation behavior changed: these are synchronous, side-effect-free config
resolvers (no `.await`, no locks, no `self`). No struct fields were widened — the
moved items are free functions and never were methods. Existing cfg behavior is
unchanged.

LoC accounting: `voice_barge_in.rs` production LoC drops 3136 -> 3044 (-92; the
97-line block leaves, offset by the +5 prod re-import comment/use). New child
content `voice_barge_in/routing.rs` grows 383 -> 487 (+104, the verbatim bodies
plus the three rustfmt-rewrapped signatures), below the 1000-line giant
threshold, so only the root ratchet is lowered.

Token proof: normalized body-token comparison against HEAD is identical after
visibility-only normalization (strip `pub(super) `, collapse whitespace). Per
moved method:

| Method | Body match | Normalized SHA |
| --- | :---: | --- |
| `agent_voice_matches_channel` | yes | `0ade5ba1c1c3` |
| `agent_voice_channel` | yes | `66d7f10defe1` |
| `agent_voice_background_channel` | yes | `c81512ba80a3` |
| `agent_voice_background_channel_for` | yes | `055f65513300` |
| `agent_voice_channel_for_background` | yes | `3746a4c15381` |
| `agent_voice_source_channel_for_background` | yes | `7cd0e5e46ebc` |
| `effective_voice_source_channel` | yes | `0b2e38020582` |
| `agent_text_channel_matches` | yes | `5d51f1ba01a3` |

REMAINING-MAP CLOSING ASSESSMENT (post-S7):

S1-S6 extracted the five S1 concerns (STT, progress playback, final-result
playback, routing resolution, live-cut playback) plus the TTS pipeline. S7 now
folds the agent-voice routing helper block — the (a) candidate the S6 review
identified — into the routing child it already serves. What still remains in the
root is:

- foreground decision/parser cluster (the S6-review (b) candidate, ~line 599+):
  `foreground_ack_text`, `generate_foreground_ack_text`, the
  `looks_like_background_work_request`/foreground-intent parsing helpers, and the
  `normalized_foreground_max_chars`/`normalized_foreground_timeout_ms`
  normalizers. These are also free, config/text-only functions with no
  struct-field dependency and are the next clean verbatim S8 candidate — NOT
  taken here to keep S7 a single coherent ownership move (the agent-voice
  routing block is its own concern; the foreground ack/parse block is a distinct
  one, so they stay as separate slices per the S-series discipline).

- routing/dispatch HOT PATH (orchestration core): the four largest methods
  (announcement 340, handoff 291, turn 244, completed-utterance 163),
  `drain_deferred_after_turn`/`take_deferred_prompt`/`apply_dispatcher_command`,
  the inflight-foreground cancel registry, the voice-context register/unregister
  surface, and `buffer_for_channel`. This remains the irreducible orchestration
  root: each large method threads SharedData, the deferred buffers, the cancel
  registry, the background driver handoff, and the extracted clusters together,
  and moving any subset verbatim would split mutually-recursive callers or force
  struct-field visibility widening (forbidden). No clean verbatim slice here.

- lifecycle/config + misc residue (constructor + public command/config surface):
  below the cohesive-cluster threshold and intentionally root-resident.

Conclusion: S7 closes the (a) agent-voice routing helper candidate. The EPIC is
NOT closed — the (b) foreground decision/parser cluster is a valid clean S8
candidate; after that only the orchestration core and the sub-threshold
lifecycle residue remain, neither of which admits a verbatim slice under the
S-series discipline (verbatim-only, zero field-visibility edits, child < 1000
LoC).

Gates run before commit:

- cargo fmt --check
- cargo check --lib
- cargo clippy --lib -- -D warnings
- set -o pipefail; cargo test --lib voice 2>&1 | grep "test result" (277
  passed) && cargo test --lib discord 2>&1 | grep "test result" | tail -1 (1636
  passed)
- python3 scripts/audit_maintainability.py --check
