+++ docs/agent-maintenance/change-surfaces.md
@@ -802,7 +802,7 @@
     was removed from `giant_file_registry.toml`; #3038 S5 locked the final
     root ratchet at 274 production lines).
   - `src/services/discord/session_runtime.rs` (1753 lines).
-  - `src/services/discord/voice_barge_in.rs` (3136 lines after #3038
+  - `src/services/discord/voice_barge_in.rs` (3044 lines after #3038
     VoiceBargeInRuntime S1 moved the STT method cluster to
     `src/services/discord/voice_barge_in/stt.rs` (314 production lines) and
     S2 moved the progress playback method cluster to
@@ -815,6 +815,8 @@
     `src/services/discord/voice_barge_in/live_cut_playback.rs` (120 production
     lines), and S6 moved the TTS pipeline cluster to
     `src/services/discord/voice_barge_in/tts_pipeline.rs` (86 production
+    lines), and S7 folded the agent-voice routing helper block into
+    `src/services/discord/voice_barge_in/routing.rs` (now 484 production
     lines);
     voice STT/TTS, lobby routing, progress mirroring, and barge-in
     orchestration surface; tracked decompose target — see
